+++ frontend/src/components/Room/hooks/index.ts
@@ -1,0 +1,5 @@
+import usePageNum from "./usePageNum"
+import usePointer from "./usePointer"
+import useZoom from "./useZoom"
+
+export { usePageNum, usePointer, useZoom }
