+++ frontend/src/components/Room/hooks/usePageNum.ts
@@ -1,0 +1,34 @@
+import { useState, Dispatch, SetStateAction } from "react"
+
+import { PageOption } from "../NavBar"
+import socket from "../../../socket"
+
+type UsePageNumReturn = [
+  number,
+  Dispatch<SetStateAction<number>>,
+  (option: PageOption) => void
+]
+
+export default (roomID: string, pages: string[]): UsePageNumReturn => {
+  const [pageNum, setPageNum] = useState(1)
+  const handleChangePage = (option: PageOption): void => {
+    const { offset, goto } = option
+    setPageNum(current => {
+      let newPageNum: number
+      if (offset) {
+        newPageNum = current + offset
+      } else if (goto) {
+        newPageNum = goto
+      }
+
+      if (newPageNum <= pages.length && newPageNum >= 1) {
+        socket.emit("client change page", { roomID, pageNum: newPageNum })
+        return newPageNum
+      }
+
+      return current
+    })
+  }
+
+  return [pageNum, setPageNum, handleChangePage]
+}
